use jsx3 and latest reason-react to see if this fixes windows ppx issues
