+++ examples/bs-7.0.1/bsconfig.json
@@ -1,7 +1,7 @@
 {
   "name": "react-template",
   "reason": {
-    "react-jsx": 2
+    "react-jsx": 3
   },
   "sources": {
     "dir" : "src",
