+++ examples/bs-7.0.1/src/App.re
@@ -1,0 +1,38 @@
+type state = {count: int};
+
+type action =
+  | Increment
+  | Decrement;
+
+let initialState = {count: 0};
+
+let reducer = (state, action) =>
+  switch (action) {
+  | Increment => {count: state.count + 1}
+  | Decrement => {count: state.count - 1}
+  };
+
+let css = _ => ();
+let label = _ => ();
+let prop = _ => ();
+let styles = [%css [prop("1"), prop("2")]];
+
+[@react.component]
+let make = () => {
+  let (state, dispatch) = React.useReducer(reducer, initialState);
+
+  <main>
+    {React.string("Simple counter with reducer")}
+    <div>
+      <button onClick={_ => dispatch(Decrement)}>
+        {React.string("Decrement")}
+      </button>
+      <span className="counter">
+        {state.count |> string_of_int |> React.string}
+      </span>
+      <button onClick={_ => dispatch(Increment)}>
+        {React.string("Increment")}
+      </button>
+    </div>
+  </main>;
+};
