--- examples/bs-7.0.1/src/Component1.re
@@ -1,29 +1,0 @@
-/* This is the basic component. */
-let component = ReasonReact.statelessComponent("Page");
-
-let css = _ => ();
-let label = _ => ();
-let prop = _ => ();
-let styles = [%css [prop("1"), prop("2")]];
-
-/* Your familiar handleClick from ReactJS. This mandatorily takes the payload,
-   then the `self` record, which contains state (none here), `handle`, `reduce`
-   and other utilities */
-let handleClick = (_event, _self) => Js.log("clicked!");
-/* let x = () => */
-
-/* `make` is the function that mandatorily takes `children` (if you want to use
-   `JSX). `message` is a named argument, which simulates ReactJS props. Usage:
-
-   `<Page message="hello" />`
-
-   Which desugars to
-
-   `ReasonReact.element(Page.make(~message="hello", [||]))` */
-let make = (~message, _children) => {
-  ...component,
-  render: self =>
-    <div onClick=(self.handle(handleClick))>
-      (ReasonReact.string(message))
-    </div>,
-};
