+++ examples/bs-7.0.1/src/Index.bs.js
@@ -1,13 +1,10 @@
 // Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
 'use strict';
 
+var React = require("react");
 var ReactDOMRe = require("reason-react/src/ReactDOMRe.js");
-var ReasonReact = require("reason-react/src/ReasonReact.js");
-var Component1$ReactTemplate = require("./Component1.bs.js");
-var Component2$ReactTemplate = require("./Component2.bs.js");
+var App$ReactTemplate = require("./App.bs.js");
 
-ReactDOMRe.renderToElementWithId(ReasonReact.element(undefined, undefined, Component1$ReactTemplate.make("Hello!", /* array */[])), "index1");
-
-ReactDOMRe.renderToElementWithId(ReasonReact.element(undefined, undefined, Component2$ReactTemplate.make("Hello!", /* array */[])), "index2");
+ReactDOMRe.renderToElementWithId(React.createElement(App$ReactTemplate.make, { }), "root");
 
 /*  Not a pure module */
