+++ examples/bs-7.0.1/src/Index.re
@@ -1,3 +1,1 @@
-ReactDOMRe.renderToElementWithId(<Component1 message="Hello!" />, "index1");
-
-ReactDOMRe.renderToElementWithId(<Component2 greeting="Hello!" />, "index2");
+ReactDOMRe.renderToElementWithId(<App />, "root");
